requirements.txt
asgiref==3.7.2
bootstrap4==0.1.0
crispy-bootstrap4==2023.1
Django==4.2.7
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-js-asset==2.1.0
django-model-utils==4.3.1
djangoajax==3.3
mysqlclient==2.2.0
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3